requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
psycopg>=3.1.14
python-dotenv>=1.0.0
SQLAlchemy>=2.0.23

# Data validation
pydantic>=2.5.1
pydantic-settings>=2.1.0
email-validator>=2.1.0.post1

# Date handling
python-dateutil>=2.8.2
pytz>=2023.3.post1

# CSV handling
pandas>=2.1.3

# API documentation
openapi-schema-pydantic>=1.2.4

# Error tracking
sentry-sdk>=1.35.0

# Logging
structlog>=23.2.0
